refactor: Split editor into different components and files
