+++ src/lib/editor/utils.ts
@@ -1,9 +1,10 @@
 import type { Project } from '$lib/projects';
 import type { User } from '$lib/user';
 import { hasPermissions } from '$lib/utils';
-import { Canvas, FabricObject, Group, Path, Point, util } from 'fabric';
+import { Canvas, Path } from 'fabric';
+import { applyObjectPermissions } from './permissions';
 
-export async function renderFromData(
+export async function loadCanvasFromData(
     canvas: Canvas,
     data: { content: any; user: User; project: Project }
 ) {
@@ -12,7 +13,11 @@ export async function renderFromData(
     if (!hasPermissions(data.user, data.project, ['allow_edit'])) {
         canvas.selection = false;
         loadedCanvas.forEachObject((object) => {
-            setPermissionsForObject(canvas, object, false, true, 'grab');
+            applyObjectPermissions(canvas, object, {
+                selectable: false,
+                bordered: true,
+                cursor: 'grab'
+            });
         });
     }
 
@@ -21,115 +26,21 @@ export async function renderFromData(
     canvas.requestRenderAll();
 }
 
-export function setPermissionsForObject(
+export function updateCanvasDimensions(
     canvas: Canvas,
-    object: FabricObject,
-    allowEdit: boolean,
-    showBorders: boolean = false,
-    cursor: string = 'default'
+    body: HTMLElement,
+    aside: HTMLElement,
+    header: HTMLElement
 ) {
-    if (showBorders === false && allowEdit === true) {
-        showBorders = true;
-    }
+    if (!canvas) return;
 
-    canvas.defaultCursor = cursor;
-    canvas.setCursor(canvas.defaultCursor);
-
-    object.lockMovementX = !allowEdit;
-    object.lockMovementY = !allowEdit;
-    object.lockRotation = !allowEdit;
-    object.lockScalingX = !allowEdit;
-    object.lockScalingY = !allowEdit;
-    object.lockSkewingX = !allowEdit;
-    object.lockSkewingY = !allowEdit;
-    object.lockScalingFlip = !allowEdit;
-    object.hasControls = allowEdit;
-    object.hasBorders = showBorders;
-    object.hoverCursor = !allowEdit ? cursor : null;
-}
-
-export function centerView(
-    canvas: Canvas,
-    data: { project: { width: number; height: number } },
-    duration: number = 500
-) {
-    const currentZoom = canvas.getZoom();
-    const targetZoom = 1;
-
-    const viewWidth = canvas.width / targetZoom;
-    const viewHeight = canvas.height / targetZoom;
-
-    const targetX = (data.project.width * 100) / 2 - viewWidth / 2;
-    const targetY = (data.project.height * 100) / 2 - viewHeight / 2;
-
-    const currentTransform = canvas.viewportTransform || [1, 0, 0, 1, 0, 0];
-    const currentX = -currentTransform[4];
-    const currentY = -currentTransform[5];
-
-    util.animate({
-        startValue: 0,
-        endValue: 1,
-        duration,
-        easing: util.ease.easeInOutCubic,
-        onChange: (value: number) => {
-            const interpolatedZoom = currentZoom + (targetZoom - currentZoom) * value;
-
-            canvas.setZoom(interpolatedZoom);
-
-            const interpolatedX = currentX + (targetX - currentX) * value;
-            const interpolatedY = currentY + (targetY - currentY) * value;
-
-            canvas.absolutePan(new Point(interpolatedX, interpolatedY));
-        },
-        onComplete: () => {
-            canvas.setZoom(targetZoom);
-            canvas.absolutePan(new Point(targetX, targetY));
-        }
+    canvas.setDimensions({
+        width: body.clientWidth - aside.clientWidth,
+        height: body.clientHeight - header.clientHeight
     });
 }
 
-export function centerViewOnObject(canvas: Canvas, object: FabricObject, duration: number = 500) {
-    const currentZoom = canvas.getZoom();
-    const targetZoom = 1;
-
-    const boundingRect = object.getBoundingRect();
-
-    const vpw = canvas.width / targetZoom;
-    const vph = canvas.height / targetZoom;
-
-    const objectCenterX = (boundingRect.left + boundingRect.width / 2) * targetZoom;
-    const objectCenterY = (boundingRect.top + boundingRect.height / 2) * targetZoom;
-
-    const targetX = objectCenterX - vpw / 2;
-    const targetY = objectCenterY - vph / 2;
-
-    const currentTransform = canvas.viewportTransform || [1, 0, 0, 1, 0, 0];
-    const currentX = -currentTransform[4];
-    const currentY = -currentTransform[5];
-
-    util.animate({
-        startValue: 0,
-        endValue: 1,
-        duration,
-        easing: util.ease.easeInOutCubic,
-        onChange: (value: number) => {
-            const interpolatedZoom = currentZoom + (targetZoom - currentZoom) * value;
-
-            canvas.setZoom(interpolatedZoom);
-
-            const interpolatedX = currentX + (targetX - currentX) * value;
-            const interpolatedY = currentY + (targetY - currentY) * value;
-
-            canvas.absolutePan(new Point(interpolatedX, interpolatedY));
-        },
-        onComplete: () => {
-            canvas.setZoom(targetZoom);
-            canvas.absolutePan(new Point(targetX, targetY));
-        }
-    });
-}
-
-function drawGrid(canvas: Canvas, data: { project: Project }) {
+export function drawGrid(canvas: Canvas, data: { project: Project }) {
     const gridSize = 100;
     const lines = [];
 
@@ -157,73 +68,4 @@ function drawGrid(canvas: Canvas, data: { project: Project }) {
 
     canvas.add(lineGroup);
     canvas.sendObjectToBack(lineGroup);
-}
-
-export interface CanvasObject {
-    object: FabricObject;
-    name: string;
-    type: string;
-    componentID?: number;
-    children?: CanvasObject[];
-}
-
-export function getCanvasObjects(canvas: Canvas): CanvasObject[] {
-    function traverseObjects(objects: FabricObject[]): CanvasObject[] {
-        return objects.map((obj) => {
-            if (obj.type === 'group' && obj instanceof Group) {
-                return {
-                    object: obj,
-                    name: 'name' in obj ? (obj.name as string) : 'Grupo',
-                    type: 'group',
-                    componentID: 'id' in obj ? (obj.id as number) : undefined,
-                    children: traverseObjects(obj.getObjects())
-                };
-            }
-
-            let name = obj.type;
-
-            if ('text' in obj) {
-                name = obj.text as string;
-            }
-
-            if ('name' in obj) {
-                name = (obj.name as string)[0].toUpperCase() + (obj.name as string).slice(1);
-            }
-
-            switch (name) {
-                case 'rect':
-                    name = 'Retângulo';
-                    break;
-                case 'circle':
-                    name = 'Circulo';
-                    break;
-                case 'path':
-                    name = 'Desenho';
-                    break;
-            }
-
-            return {
-                object: obj,
-                name: name,
-                type: obj.type
-            };
-        });
-    }
-
-    const objects = canvas?.getObjects().filter((x) => x.selectable && x.evented) ?? [];
-    return traverseObjects(objects);
-}
-
-export function updateCanvasDimensions(
-    canvas: Canvas,
-    body: HTMLElement,
-    aside: HTMLElement,
-    header: HTMLElement
-) {
-    if (!canvas) return;
-
-    canvas.setDimensions({
-        width: body.clientWidth - aside.clientWidth,
-        height: body.clientHeight - header.clientHeight
-    });
 }
